vehicle-software-simulator/requirements.txt
python-can
Flask
flask-cors
websockets
